Add validation of FeedEntities
If an update makes the cache inconsistent so that a departure from an
earlier stop is predicted to happen at the same time or later than an
arrival from a later stop, mark the FeedEntity as invalid with a new
property "isValid" set to false. The next cache update might change the
validity again.